pyproject.toml
[project]
name = "unitree_webrtc_connect"
version = "2.0.4"
description = "Connect to the Unitree Go2 and G1 with WebRTC"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "legion1581", email = "legion1581@gmail.com" }
]

dependencies = [
    "aiortc>=1.9.0",
    "pycryptodome",
    "requests",
    "flask-socketio",
    "lz4",
    "numpy",
    "packaging",
    "sounddevice",
    "pyaudio",
    "pydub",
    "opencv-python"
]


[build-system]
requires = ["uv_build>=0.10.9,<0.11.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-root = ""
